workshop-3/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
transformers
pillow
python-multipart
pydantic
